chore: FeatureInterface 제거
