+++ Projects/Feature/Home/HomeFeature/Sources/HomeReducer.swift
@@ -6,14 +6,28 @@
 //  Created by JiYeon
 //
 
-import HomeFeatureInterface
 import ComposableArchitecture
 
-extension HomeReducer {
-  public static let HomeReducer = Reduce<State, Action> { state, action in
-    switch action {
-    default:
-      return .none
+@Reducer
+public struct HomeReducer {
+  public init() {}
+  
+  @ObservableState
+  public struct State: Equatable {
+    
+    public init() {}
+  }
+
+  public enum Action: BindableAction, Equatable {
+    case binding(BindingAction<State>)
+  }
+
+  public var body: some ReducerOf<Self> {
+    BindingReducer()
+    Reduce { state, action in
+      switch action {
+      default: return .none
+      }
     }
   }
 }
