--- Projects/Feature/Home/HomeFeatureInterface/Sources/HomeReducer.swift
@@ -1,33 +1,0 @@
-//
-//  HomeReducer.swift
-//
-//  Home
-//
-//  Created by JiYeon
-//
-
-import ComposableArchitecture
-
-@Reducer
-public struct HomeReducer {
-  private let reducer: Reduce<State, Action>
-  
-  public init(reducer: Reduce<State, Action>) {
-    self.reducer = reducer
-  }
-  
-  @ObservableState
-  public struct State: Equatable {
-    
-    public init() {}
-  }
-
-  public enum Action: BindableAction, Equatable {
-    case binding(BindingAction<State>)
-  }
-
-  public var body: some ReducerOf<Self> {
-    BindingReducer()
-    reducer
-  }
-}
