--- Projects/Feature/Home/HomeFeatureInterface/Sources/Map/MapViewRepresentable.swift
@@ -1,9 +1,0 @@
-//
-//  MapViewRepresentable.swift
-//  HomeFeature
-//
-//  Created by Jiyeon on 6/7/25.
-//  Copyright © 2025 Sseudam.a2bo.ios. All rights reserved.
-//
-
-import Foundation
