+++ Projects/Sseudam/Project.swift
@@ -24,7 +24,7 @@ let project = Project.makeApp(
     ],
   ],
   dependencies: [
-    .Features.Home.Implement,
+    .Features.Home,
     .Domain.Home.Implement,
     .Data.Home.Implement
   ]
