+++ Projects/Sseudam/Sources/SseudamView.swift
@@ -8,6 +8,7 @@
 
 import SwiftUI
 import ComposableArchitecture
+import HomeFeature
 
 struct SseudamView: View {
   @Bindable var store: StoreOf<SseudamFeature> = Store(
@@ -20,7 +21,7 @@ struct SseudamView: View {
     VStack {
       switch store.selectedTab {
       case .home:
-        EmptyView()
+        HomeView(store: store.scope(state: \.home, action: \.home))
       case .search:
         EmptyView()
       case .profile:
@@ -38,4 +39,5 @@ struct SseudamView: View {
 #Preview {
   SseudamView()
 }
+
